requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.2
Flask-Caching>=2.0.2
Flask-Limiter>=3.4.0
Flask-Talisman>=1.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
requests>=2.28.0
schedule>=1.1.0
gunicorn>=20.1.0
redis>=4.0.0
pytest>=7.0.0
Werkzeug>=2.0.0
